finished reusable table
